Docs changes from PR feedback
