+++ docs/src/pages/ui/primitives/textfield/react.mdx
@@ -39,7 +39,7 @@ import '@aws-amplify/ui-react/styles.css';
 
 ### Accessibility / Label behavior
 
-The TextField primitive is accessible by default. A matching `label` HTML element is rendered with the `TextField` by default,
+The TextField primitive is accessible by default. A matching `label` HTML element is rendered with the `TextField`,
 all that is required is to provide a `label` prop with a `string` or `ReactNode`. If no `id` is provided, one will be automatically
 generated and connected to both `label` and `input` elements.
